Improve compatibility and precision of date information in log filenames
There a few reasons that the current log filenames may not be ideal:

1) They contain `:` which breaks in Windows and is shown as `/` in OSX GUI ([#144](https://github.com/ericaltendorf/plotman/issues/144))
2) They don't contain any more precision than seconds, so with many concurrent plots being created (or error situations), naming collisions can occur
3) They don't contain any timezone information which makes them not fully comprehensive

This potentially closes https://github.com/ericaltendorf/plotman/issues/144

Note: As discussed on chat with @altendky, we settled on `+HH_MM` / `-HH_MM` for the timezone part (`±[hh]:[mm]` with the same `:` -> `_` done for times).  it seems that `strftime` provides the offset as one component via `%z`, without the option to use a `HH:MM` format.  For now I have left it without a separator, but it would be trivial to insert a `_` if it is desired. Thoughts?